port libmesode-specific API
This patch syncs this library with the additional libmesode functionality.

Instead of a different API signature for `xmpp_connect_client()` we
provide an API function `xmpp_conn_set_certfail_handler()` which enables
the new functionality.

Instead of module-global variables in `tls_openssl.c` we attach the
data to the `SSL_CTX` and retrieve it when required.

All functions previously returning `char*` where ownership isn't
transferred return now a `const char*`.